fixed issues with spinning wheel after login
