+++ src/context/UserContext/index.ts
@@ -1,2 +1,3 @@
 export * from './UserContext'
+export * from './UserProvider'
 export { useUserContext } from './useUserContext'
